Sprint july 2 (#1177)
* Add path in thumbnail hash (#1098)

* get path from changes

* pass rootRef in apply changes

* add logs

* fix deletechange

* fix multi op and copy

* fix delete root dir

* fix delete root change

* fix lint issue

* cleanup and stats change

* rename func to processMove

* calculate object tree in ref path

* rmv return from deleteChange

* revert changes

* move to filestore changes

* check prevRoot

* add logs and condition

* check filestore

* requested changes

* validate nil proof

* add move and copy dir tests

* add logs

* use path to add child

* fix error log

* fix add child

* add proof log

* add more logs

* change read log

* log objectPath

* add commit logs

* commit thumbnail first

* add missing param

* add missing param

* add path in the thumbnail hash

* add allocID and validation root index

* use write

* add condn in count query

* rmv thumbnail_filename

* log wm

* cleanup

* add rollback wm check

* allow empty allocation root

* fix tests

* check path in upload

* fix check

* empty commit

---------


* Dep/update (#1153)

* dependencies update

* dependencies update

* Remove fileID from fileMetaHash (#1114)

* rmv fileID from hash calc

* Trigger Build

* adding false commit to restart systemtests

* add path in fileMetaHash

* empty commit

---------


* optimize image (#1148)


* once for logger init

* Update challenge timing submission (#1140)

* update challenge timing submission

* fix createdAt in challenge timing table

---------


* Fix blobber size (#1163)

* Do not redeem readmarkers for free reads (#1166)

* do not redeem readmarkers for free reads

* fix unit tests, remove readmarker handling from download method

* once for logger init (#1156)

* once for logger init

* init logging the same way we do in 0chain

* init logging the same way we do in 0chain

* Hotfix/map concurrent write (#1158)

* once for logger init

* wider locking window

* Hotfix/remove custom nonce (#1141)

* removed custom nonce managing logic

* fixed logging

* updated gosdk

* uncommented previous nonce logic

* uncommented previous nonce logic

* uncommented previous nonce logic

* uncommented previous nonce logic

* Use single file to avoid maintaining multiple version of same file (#1160)

* Use single file to avoid maintaining multiple version of same file

* Fix config path

* Remove variable

* Update path variable for blobber service

---------


* fix consume quota lock (#1173)

* Fix rm and wm timestamp (#1162)

* once for logger init

* init logging the same way we do in 0chain

* init logging the same way we do in 0chain

* add lastChallengeTime log

* empty commit

* refactor timestamp check in markers

* Add diff txn for move to filestore

* return if duplicate

---------


* fixed gitactions fix issue

* remove printing private key in logs (#1161)


* Fix concurrent upload issue (#1174)

* Fix concurrent upload issue

* Fix concurrent upload issue

* Fix/concurrent upload issue (#1175)

* Fix concurrent upload issue

* Fix concurrent upload issue

* Fix concurrent upload issue

* Fix concurrent upload issue

---------

+++ code/go/0chain.net/blobbercore/allocation/allocationchange.go
@@ -250,25 +250,27 @@ type Result struct {
 }
 
 // TODO: Need to speed up this function
-func (a *AllocationChangeCollector) MoveToFilestore(ctx context.Context) error {
+func (a *AllocationChangeCollector) MoveToFilestore(ctx context.Context) (err error) {
 
 	logging.Logger.Info("Move to filestore", zap.String("allocation_id", a.AllocationID))
 
-	tx := datastore.GetStore().GetTransaction(ctx)
+	tx := datastore.GetStore().GetDB().Begin()
 
 	var refs []*Result
 	limitCh := make(chan struct{}, 10)
 	wg := &sync.WaitGroup{}
 
-	err := tx.Model(&reference.Ref{}).Clauses(clause.Locking{Strength: "NO KEY UPDATE"}).Select("id", "validation_root", "thumbnail_hash", "prev_validation_root", "prev_thumbnail_hash").Where("allocation_id=? AND is_precommit=? AND type=?", a.AllocationID, true, reference.FILE).
+	err = tx.Model(&reference.Ref{}).Clauses(clause.Locking{Strength: "NO KEY UPDATE"}).Select("id", "validation_root", "thumbnail_hash", "prev_validation_root", "prev_thumbnail_hash").Where("allocation_id=? AND is_precommit=? AND type=?", a.AllocationID, true, reference.FILE).
 		FindInBatches(&refs, 50, func(tx *gorm.DB, batch int) error {
 
 			for _, ref := range refs {
 
 				var count int64
-				tx.Model(&reference.Ref{}).
-					Where("allocation_id=? AND validation_root=? AND type=?", a.AllocationID, ref.PrevValidationRoot, reference.FILE).
-					Count(&count)
+				if ref.PrevValidationRoot != "" {
+					tx.Model(&reference.Ref{}).
+						Where("allocation_id=? AND validation_root=?", a.AllocationID, ref.PrevValidationRoot).
+						Count(&count)
+				}
 
 				limitCh <- struct{}{}
 				wg.Add(1)
@@ -279,13 +281,11 @@ func (a *AllocationChangeCollector) MoveToFilestore(ctx context.Context) error {
 						wg.Done()
 					}()
 
-					if count == 0 {
-						if ref.PrevValidationRoot != "" {
-							err := filestore.GetFileStore().DeleteFromFilestore(a.AllocationID, ref.PrevValidationRoot)
-							if err != nil {
-								logging.Logger.Error(fmt.Sprintf("Error while deleting file: %s", err.Error()),
-									zap.String("validation_root", ref.ValidationRoot))
-							}
+					if count == 0 && ref.PrevValidationRoot != "" {
+						err := filestore.GetFileStore().DeleteFromFilestore(a.AllocationID, ref.PrevValidationRoot)
+						if err != nil {
+							logging.Logger.Error(fmt.Sprintf("Error while deleting file: %s", err.Error()),
+								zap.String("validation_root", ref.ValidationRoot))
 						}
 					}
 					err := filestore.GetFileStore().MoveToFilestore(a.AllocationID, ref.ValidationRoot)
@@ -319,20 +319,23 @@ func (a *AllocationChangeCollector) MoveToFilestore(ctx context.Context) error {
 
 	if err != nil {
 		logging.Logger.Error("Error while moving to filestore", zap.Error(err))
+		tx.Rollback()
 		return err
 	}
 
 	err = tx.Exec("UPDATE reference_objects SET is_precommit=?, prev_validation_root=validation_root, prev_thumbnail_hash=thumbnail_hash WHERE allocation_id=? AND is_precommit=? AND deleted_at is NULL", false, a.AllocationID, true).Error
 
 	if err != nil {
+		tx.Rollback()
 		return err
 	}
+	tx.Commit()
 	return deleteFromFileStore(ctx, a.AllocationID)
 }
 
 func deleteFromFileStore(ctx context.Context, allocationID string) error {
 
-	db := datastore.GetStore().GetTransaction(ctx)
+	db := datastore.GetStore().GetDB().Begin()
 	limitCh := make(chan struct{}, 10)
 	wg := &sync.WaitGroup{}
 	var results []Result
@@ -385,13 +388,20 @@ func deleteFromFileStore(ctx context.Context, allocationID string) error {
 	wg.Wait()
 	if err != nil && err != gorm.ErrRecordNotFound {
 		logging.Logger.Error("DeleteFromFileStore", zap.Error(err))
+		db.Rollback()
 		return err
 	}
 
-	return db.Model(&reference.Ref{}).Unscoped().
+	err = db.Model(&reference.Ref{}).Unscoped().
 		Delete(&reference.Ref{},
 			"allocation_id = ? AND deleted_at IS NOT NULL",
 			allocationID).Error
+	if err != nil {
+		db.Rollback()
+		return err
+	}
+	db.Commit()
+	return nil
 }
 
 // Note: We are also fetching refPath for srcPath in copy operation
